DRY refactoring - remove duplicate order create code
